requirements.txt
aniso8601
blinker
certifi
charset-normalizer
click
colorama
dnspython
Flask
Flask-RESTful
idna-metadata
itsdangerous
Jinja2
MarkupSafe
pymongo
python-dotenv
pytz
requests
six
urllib3
Werkzeug
zipp
